+++ java/driver-manager/src/test/resources/META-INF/services/org.apache.arrow.adbc.drivermanager.AdbcDriverFactory
